requirements.txt
Package             Version
------------------- -------
asgiref             3.5.0
atomicwrites        1.4.0
attrs               21.4.0
colorama            0.4.4
Deprecated          1.2.13
Django              3.2.12
django-redis        5.2.0
djangorestframework 3.13.1
importlib-metadata  4.11.1
iniconfig           1.1.1
mock                4.0.3
packaging           21.3
pip                 22.0.3
pluggy              1.0.0
psycopg2            2.9.3
py                  1.11.0
pyparsing           3.0.7
pytest              7.0.1
pytest-django       4.5.2
pytz                2021.3
redis               4.1.3
setuptools          47.1.0
sqlparse            0.4.2
tomli               2.0.1
typing_extensions   4.1.1
wrapt               1.13.3
zipp                3.7.0
